+++ exonum/src/blockchain/schema.rs
@@ -16,11 +16,11 @@ use super::{config::StoredConfiguration, Block, BlockProof, Blockchain, Transact
 use crypto::{CryptoHash, Hash, PublicKey};
 use helpers::{Height, Round};
 use messages::{Precommit, RawMessage};
+use node::ConnectInfo;
 use storage::{
     Entry, Fork, KeySetIndex, ListIndex, MapIndex, MapProof, ProofListIndex, ProofMapIndex,
     Snapshot,
 };
-use node::ConnectInfo;
 
 /// Defines `&str` constants with given name and value.
 macro_rules! define_names {