+++ exonum/src/encoding/serialize/json.rs
@@ -273,8 +273,10 @@ impl ExonumJson for SocketAddr {
 }
 
 impl ExonumJsonDeserialize for SocketAddr {
-    fn deserialize(value: &Value) -> Result<Self, Box<dyn Error>> where
-        Self: Sized {
+    fn deserialize(value: &Value) -> Result<Self, Box<dyn Error>>
+    where
+        Self: Sized,
+    {
         serde_json::from_value(value.clone()).map_err(|e| e.into())
     }
 }