+++ exonum/src/events/codec.rs
@@ -151,9 +151,14 @@ mod test {
 
     fn create_encrypted_codecs() -> (MessagesCodec, MessagesCodec) {
         let (public_key, secret_key) = gen_keypair_from_seed(&Seed::new([1; SEED_LENGTH]));
-        let addr: SocketAddr = "127.0.0.1:8000".parse().unwrap();
-        let mut params =
-            HandshakeParams::new(public_key, secret_key, SharedConnectList::default(), 1024, addr);
+        let address: SocketAddr = "127.0.0.1:8000".parse().unwrap();
+        let mut params = HandshakeParams::new(
+            public_key,
+            secret_key,
+            SharedConnectList::default(),
+            1024,
+            address,
+        );
         params.set_remote_key(public_key);
 
         let mut initiator = NoiseWrapper::initiator(&params).session;